Move yaml handler into util subdirectory
